ADD добавлено преобразование Мощность - Время - Энергия

+++ e-calc/src/PhysicalQuantity/QuantityTemperature.cs
@@ -14,7 +14,7 @@ namespace e_calc
             Units = new List<Unit>()
             {
                 new Unit(   "ºC",   UnitsEnum.C,   1  , 0),
-                new Unit(   "ºF",   UnitsEnum.F, (double) 5/9, (double) -32*5/9),
+                new Unit(   "ºF",   UnitsEnum.F, 5.0/9.0, -32.0*5.0/9.0),
                 new Unit(   "К",    UnitsEnum.K,  1, -273.15)
             };
         }
